Add a smoke test for didChange requests
Summary:
For most requests, we have some basic unit test coverage to make
sure the method executes, and that telemetry is hooked up correctly.

We don't have this for didChange, probably because we just manually
verified the telemetry when we added it. But I'm working on adding
alpha-quality type-errors-for-unsaved-changes (that will be in a
stacked diff), and it's nice ot have a basic test harness in place
before I do that.

Reviewed By: jasontatton-meta

Differential Revision: D42685094

fbshipit-source-id: f9816648b596052712aee49d22f85914af53391a