Fix flask model query, round 2
Summary:
This is still broken, because the actual fully qualified name for a local decorator is something like `$module_name$app.route`, which doesn't match the regular expression.

In a follow-up, we will try to remove the `$module$` prefix when evaluating constraints, since this is an implementation detail which shouldn't matter for a user.

Reviewed By: alexkassil

Differential Revision: D42791677

fbshipit-source-id: 330697a2a407710322f54bfb9a717200c56630af