Back Pyre off to Ocaml 4.14.0
Summary:
After more extensive investigation the Hack team has
realized that the current logic in Ocamlrep is tied to the Ocaml 4
runtime in ways that will take time to migrate.

Unfortunately, this means Pyre can only have one in the short term:
Ocamlrep-based Errpy integration or Ocaml 5.0. Given that Errpy is
an immeidate blocker for IDE work, whereas Ocaml 5.0 is mostly a
BE win and a potential blocker for perf problems, the right
move for now is probably to back off.

The good news is that there was a great deal of other work
(library upgrades, fixing use of deprecated names in the standard libary,
etc) that we had to do to be Ocaml 5.0 ready, and that is all done.

So if multiprocessing / serialization - related perf overhead become
a top concern, we can either do a sprint on ocamlrep or find an
alternative (e.g. wrap Errpy in C and use a hand-written binder like
Jia did for CPython) and move back to 5.0

Differential Revision: D42805249

fbshipit-source-id: 07e112a57bbc2caa9c508724d7b95e247f03c52e