Add python-sideplaceholder for incremental update telemetry
Summary:
To support tracking Pyre's performance, I want the frontend language server
to have access to information about how long incremental updates take.

As a result, I want to add a new subscription message where the server
can tell us how long an incremental check took from getting the watchman
message to actually returning results (and eventually also probably subspans;
I'd like to use a map for the data so that it's not too much work to slice
the duration more granularly in the future).

Since we sometimes roll out the client and binary separately, we need to first
ship placeholder code so that when clients encounter the new telemetry event
they do not throw errors.

Reviewed By: connernilsen

Differential Revision: D42815041

fbshipit-source-id: f64ee59bf0fabaf50b1303945fb56210e043754f