Rename subscription event handlers
Summary:
I distinctly remember these names confusing me when I first started
working on the LSP: the name makes it sound like each function is creating
a subscription, which is a plausible idea (as a n00b, it's not necessarily
obvious that there's just one subscription, in theory we could use different
ones for different types of messages).

But that's not what's going on at all, these methods are not subscription
handlers but *event* handlers - the main subscribe loop parses events and
dispatches to these handlers.

So, `s/subscription/event` in the name makes more sense.

Reviewed By: grievejia

Differential Revision: D42863178

fbshipit-source-id: 0eb8c7a41d6e290f043df074b2348278d656d4d8