Move daemon restart logic up
Summary:
We have code for restarting the background task, which in turn ensures are
daemon connection is up, that we run on specific requests.

Currently, this logic is embedded in the request handlers for each specific
request. This diff moves that handling up to the dispatch.

The immediate benefit of doing this is that anyone trying to audit when exactly
we attempt to restart can more easily see this, since they have to skim just
one function that fits on a single screen.

More importantly, though, this will enable us to break the dispatcher off into
a separate foreground task, which is at the heart of a refactor that I want
both for better telemetry handling and for dealing with type errors on unsaved
changes.

Reviewed By: grievejia

Differential Revision: D42850506

fbshipit-source-id: a977b798c051f70dc75c6b4c52b3d02a2dcaf722