Split api vs dispatcher
Summary:
This commit is at the heart of a refactor I've been working
toward in this stack: split two layers of responsibility
that until now were both handled by PyreLanguageServer
into separate layers:

- PyreLanguageServerDispatcher is the foreground task
  - It's primary responsibility is to provide a loop for
    dealing with events from standard input.
  - To do that requires handling the following:
    - parsing input requests
    - deciding when to

-

----------------

The name PyreLanguageServerDispatcher is a bit of a mouthful, and it would
be reasonable to just call this class the PyreLanugageServer given that it is the
top-level logic with an async `run` method invoked by the entrypoints.

Why did I decide on the more verbose name?

I want to have good vocabulary for our architecture we need to distinguish the
dispatcher and the api clearly, and since the api is also language server
related I think calling them PyreLanguageServer{Api,Dispatcher} is clearer than
having one be called `PyreLanguageServer` and the other
`PyreLanguageServerApi`, which would invite confusion if people (and/or stale
variable names) referred to the api as a "server".

Reviewed By: grievejia

Differential Revision: D42855410

fbshipit-source-id: 705ec853d67a7c7ae55740947f81fddece0495ab