Move wait_for_exit logic up
Summary:
In the current Pyre code, the `process_.*` handler functions are
all similar - they take a typed request as input and deal with
responding to the client, returning None - except for the
`process_shutdown_request` method.

That method returns an exit code, and moreover has to watch
standard input as it waits for the "exit" request that is
supposed to follow confirmation of a "shutdown" request.

This commit extracts that shutdown code to be called from inside
the dispatcher. This will be important in an upcoming request where
I split dispatching into its own class, because `wait_for_exit`
needs access to an input channel so that it can parse the exit
request, and I want the layer with `process_.*` functions to never
directly read input.

Reviewed By: grievejia

Differential Revision: D42854242

fbshipit-source-id: 2cbf48dc332345c4bf4f54d07ba0f6b621440a76