Improve naming in the querier tests
Summary:
THe name of the test suite was RequestHandlerTest, which was
a nod toward the old name of the querier (I renamed it because
"request handler" could mean many different things, whereas this
component exists merely to execute queries against the daemon over
a socket, hence "querier").

In addition, the variable actually used for the system under test
was "pyre_query_manager", which was based on an old queue-based
implementation that we changed almost 6 months ago.

Update all the names. There is no change here other than names.

Reviewed By: grievejia

Differential Revision: D42861189

fbshipit-source-id: afdb20c4621dc4b3f7e7a36f10e83fad35dc5fa2