s/handle_request/dispatch_request
Summary:
In our current naming shemes, the top-level code for dispatching
requests is called the "handler", and each of the individual methods
to deal with a single request use the naming convention
`process_...`.

This naming scheme is consistent, but in actual discussions we
rarely refer to the layer of  dealing with requests as the "processing" layer,
that terminology is too generic to mean much - we often call these methods
"handlers" informally.

As a result, I'd like to more clearly emphasize in the naming that the top level
handler is a "dispatcher", i.e. a function whose sole responsibility is to dispatch
requests to various lower-level handling code by switching on the method,
so that the lower-level functions can be referred to as "handlers" without confusion.

Reviewed By: grievejia

Differential Revision: D42850616

fbshipit-source-id: fdd61ca5058b9aa8be597165b98ce578a5c3630c