Split up PersistentTest class
Summary:
For quite some time, I've found it tricky to skim the persistent_test
module. One of the reasons is that the test classes are too large
to serve as effective namespaces: we often have 3-4 tests of various
specific bits of logic in chunks, but we don't split them into separate
classes.

This not only makes the code harder to skim, it makes the output of test
reporting less useful, because the class portion of the test namespacing
doesn't have much value when the test classes are too big.

As a result, I'm going to work on splitting the biggest test suites into
parts. This commit starts with PersistentTest, which was a catch-all for
testing top-level helper functions. Since it only really tests two helpers,
let's just make a test suite for each one: reading requests vs initializing.

Reviewed By: grievejia

Differential Revision: D42860038

fbshipit-source-id: 0c3997cd996568b8651131ce3f852e0c6e12434e