De-abstract error event handling
Summary:
There is no reason why we need an abstract method for handling
error events sent over the daemon connection; the logic is identical
for codenav and classic servers and is unlikely to change.

It's true that in principle the shape of the message could differ between
backends, but we're not abstracted over that anyway because
we pinned the type to be `subscription.Error`, so we gain nothing by
using an abstract method.

Reviewed By: grievejia

Differential Revision: D42863177

fbshipit-source-id: 23aed576ff61ab54753f7ab07b14c31a47c5ce21