Only log DISCONNECTED if we hit an exception in `run`
Summary:
I'm beginning to prepare for a refactor where I break up
launch_and_subscribe_handler to isolate the basic control flow
as much as possible from the business logic of mutating state, etc.

But before I make pure-refactoring changes, I need to make sure the
logic actually makes sense. In the case of the `finally` block in `run`,
I think this is not true.

My reasoning is as follows:
- you can audit every branch of control flowin `launch_and_subscribe`:
  - two branches (the initial connect_and_subscribe with no connection
    failure, and a connection failure leading to a StartSuccess) lead
    to us calling into subscribe, which is a no-return loop. The only
    way out is an exception, and that exception triggers our subscriber
    to die (this is true even if the daemon is up, because we use a
    context manager for the socket connection so e.g. an uncaught python
    error will still disconnect)
  - *all* other branches return None, and they correspond to some kind of
    startup failure condition, and every single branch logs an lsp event
    (either NOT_CONNECTED if the failure is retriable or SUSPENDED if it
    is)

So, the only way we can hit the `finally` block without hitting one of the
exception handlers is if we never actually connected, and instead startup
failed - in which case we've already logged either NOT_CONNECTED or
SUSPENDED.

As a result, I'm pretty sure it's a bug that we are logging DISCONNECTED
here. We didn't actually disconnect, and moreover we are logging duplicate
events. I can't prove this to be true because there's no spec saying what's
actually intended, but for any realistic use of telemetry I think the
current code is wrong.

Hence this diff, which is a minimal diff to make the finally block
unnecessary. I'll stack a diff that moves the logging + status handling
into a helper method, which eventually will be extracted into a different
class, but the messy code here makes the semantic change clearer.

Reviewed By: connernilsen

Differential Revision: D42876661

fbshipit-source-id: 4139e57e4ebe4f74867f0f9e8516f49dda9d283f