Remove server options from initialization code
Summary:
The async_try_initialize_loop function (and the async_try_initialize
function it calls) already have injectable functionality to map an
initialization parameters (i.e. request) object into an initialization
result (i.e. response).

But that code assumes we want to take a pyre_server_options object as
input, which makes it not as flexible as it could be. Let's just
close over that in the callback we're already passing instead of locking
down the interface unnecessarily.

Reviewed By: grievejia

Differential Revision: D42911603

fbshipit-source-id: dffba3b1a2b574907558910c2bab3c10d71603f2