Add daemon_launcher class
Summary:
Background:
This diff is the first diff in a stack focused on improving the usability of our CodeNavConnection API. We want the ability to query pyre servers programmatically in use cases like linting. This functionality exists, but the semantics around the pyre server are fuzzy. The biggest problem currently is the fact that the codenav server has to be manually started before querying with a connection.

Contents:
- A base class for an API that will allow for starting and stopping the codenav server. It will deal directly with initialization.async_start_pyre_server and skip the `PyreDaemonLaunchAndSubscribeHandler` logic.

Word of caution: This API is async and will block the program (linter) until the server is fully started up.

*Why not use PyreDaemonLaunchAndSubscribeHandler?*
`PyreDaemonLaunchAndSubscribeHandler` does more than we would like since its meant to be run in a background green thread. Since we are trying to synchronously start the server before other API calls, we don't want that feature.

*Why not split `PyreDaemonLaunchAndSubscribeHandler` into `PyreDaemonLaunchHandler` and `PyreDaemonSubscribeHandler`?*
The only method we really need from what would be `PyreDaemonLaunchHandler` is `initialization.async_start_pyre_server` which we already have access to.

I'd like to get feedback early on the location of this module + the abstraction boundary: do we want to dive more into the specifics of a server start failure? Such as `BuckStartFailure` etc? (I don't think it matters yet) Or more details on a stop? (I'm not sure about stop since it seems like we don't care as much if it doesn't stop and we probably have [less information](https://fburl.com/code/7jf9kazg))

Reviewed By: grievejia

Differential Revision: D42806324

fbshipit-source-id: 76152a14c04e81ec3b76508a6e353109c24b2c2c