Handle global leak detection for If statement
Summary:
This diff stack implements a global leak detection framework on top of Pyre.

We would like the ability to check for global leaks that occur in If statements. Specifically, we need to add support for conditions of `If`, statements within `If` blocks will be handled by `forward`. The assert case of `If` is not implemented, and therefore will not catch global leaks that occur in the condition.  We call `forward_assert` for those cases.

Reviewed By: connernilsen

Differential Revision: D43045248

fbshipit-source-id: 8641ec14c800aef7e7af75b18beb5252bbf3ef58