Handle global leak detection for For statement
Summary:
This diff stack implements a global leak detection framework on top of Pyre.

We would like the ability to check for global leaks that occur in `For` statements. We dont need to do anything additional for this statement case. The statements within `For` are handled by `forward` and the iterator is handled by `Assign`. This diff adds various test cases for the `For` statement case.

Reviewed By: connernilsen

Differential Revision: D43045246

fbshipit-source-id: db179b64c9add7478a77cdd6082bc4d867a40df8