Handle global leak detection for while statement
Summary:
This diff stack implements a global leak detection framework on top of Pyre.

We would like the ability to check for global leaks that occur in `While` statements. Specifically, we need to add support for conditions of `While`, statements within `While` blocks will be handled by `forward`. The assert case of is already implemented, and therefore will catch global leaks that occur in the condition.  We call `forward_assert` for those cases. Since errors from the `[assert (not foo)]` in the else-branch is same `[foo]` as in the true-branch we can ignore the false branch.

Reviewed By: connernilsen

Differential Revision: D43045247

fbshipit-source-id: 05a627b7cce547dd760704b3a103328d484949e3