Handle global leak detection for With statement
Summary:
This diff stack implements a global leak detection framework on top of Pyre.

We would like the ability to check for global leaks that occur in `with` statements. We dont need to do anything additional for this statement case. The statements within `with` are handled by `forward` and the items of with is handled by `Assign`. This diff adds various test cases for `with`.

Reviewed By: connernilsen

Differential Revision: D43079622

fbshipit-source-id: ab36b35b480637d57916c4ab8b22b9944e0705ab