Add a since-last-ready timer and use it in telemetry
Summary:
This commit adds, to each pyre IDE request handler, the logic to include
how long the server state has been anything *other* than READY as of the
start request handling.

Why do we want this?  For the sake of tracking Pyre IDE uptime (particularly in
ML projects, but really across all of them), we're trying to track server
status in order to get a per-active-minute availability metric for Pyre.
But after building an initial server-status-per-minute metric, what we saw
was that about half of minutes involve an incremental check.

This is inconvenient for getting simple uptime metrics because in itself an
incremental check is expected, and we can't know purely from that status
whether it is a problem. Moreover, in the current server we can't use
duration + status because didChange events don't currently block, so the
durations wouldn't be useful.

The solution here is to track, in addition to duration, the amount of
time Pyre has been out of the READY state *prior* to our request. Why
make this only prior to our request rather than inclusive of the duration
of the request itself? Two reasons
- Principled: it's usually nicer for analytics and dashboarding to have
  mutually exclusive metrics, so givent that the request latency is covered by
  duration already I would prefer that this metric not double-count.
- Pragmatic: the actual implementation would be trickier if we wanted to get
  time at the end of requests, because we'd have to deal with the fact
  that the old timer can disappear

Reviewed By: grievejia

Differential Revision: D43064002

fbshipit-source-id: c66d3789b38135e6ac9c6a52d5d31c59e5e9032a