Add integration tests for generating triggered sinks from format strings
Summary:
Facebok:
# Context
We decided to use the support of multi-source rule to detect more flows into
implicit strings (that are introduced onto expressions, based on their runtime
values).
To improve the user experience, we also decided to auto-generate rules, models,
sources and sinks, whenever the above approach is used, rather than requiring
the users to maually specify these.

# This diff stack
Format strings are a common place where sources and implicit sinks meet. For
example, many URLs (which are implicit sinks) are constructed with format
strings. Hence, ideally we need to specify models for format strings to detect
such flows.  However, since it is difficult to write models for format strings,
we will instead hardcode the analysis for format strings.

# This diff
This diff specifies the goal of the above hardcoded analysis of format strings.
That is, whenever a format string contains a string literal or an expression
that has an implicit sink, we should introduce a triggered sink on all other
expressions used in the format string, such that the triggered sink can be
propagated up.

Reviewed By: alexkassil

Differential Revision: D43010126

fbshipit-source-id: 44c19ab8746389f563e528076590030f5ab9f17b