Improve the parsing of combined source rules
Summary:
The result of function `parse_combined_source_rules` is a tuple of 3 elements,
the handling of which is not very readable (e.g., we currently use a
`List.unzip3` to extract the elements from the tuple). It is better to define a
module to store such information.

This refactor will also benefit the later diffs, which will reuse
`parse_combined_source_rules` for parsing new rules that detect flows into
implicit sinks.

Reviewed By: arthaud

Differential Revision: D43092402

fbshipit-source-id: e214f365d358e3a29346355fe4d477d82c07ca12