Use `None`, not `"None"`, as error message if no error
Summary:
At one point we set the error message to be the string "None" when
there is no error. I think this was driven by a notion that having
str type is simpler than Optional[str], but in reality in the downstream
queries it is more idiomatic to use a NULL check.

More importantly, other telemetry events are using nulls, so having
a string "None" in just one kind of event is inconsistent and complicates
things.

Reviewed By: grievejia

Differential Revision: D43165068

fbshipit-source-id: d5e713654e9dea0d77d3eeb16bdaea67de98f911