Remove status after data from telemetry
Summary:
The status after isn't very informative. On a blocking operation, we will almost
always transition to READY whereas on a nonblocking operation we will almost always
keep the same status.

As a result, there's no reason for us to be recording this, it just adds line noise
and increases our message size.

Reviewed By: kinto0

Differential Revision: D43161153

fbshipit-source-id: 53e27e0c2672f58928acf18c6bcc4836dd8fa636