Add telemetry to didSave events
Summary:
Record the status prior to a didSave event and the time since last ready. We can
use these as part of our uptime OKR; it doesn't matter all that much since didChange
is probably good enough but it's a bit strange to be missing this data for just
didSave events.

I set the duration_ms to always be zero for now because we never execute a query
so the boilerplate of setting up a timer would be line noise, but it's better not
to have NULL events in our telemetry tables.

Reviewed By: kinto0

Differential Revision: D43161044

fbshipit-source-id: 81da271785142e2fed704b68ca295b83146ce148