Rename DaemonStatus to DaemonStatusTracker
Summary:
Two reasons for the rename: the first is that this is a mutable class with internal
state, so calling it a tracker makes the code more self-describing.

The second is that I'm intending to make a DeamonStatus dataclass that can encapsulate
more status information than just an enum (for now, it's going to be the enum plus the
time since last ready; in theory we could add more if needed for perf analysis). Doing
this will make the code easier to read and more bug resistant (one of the problems
with our current telemetry is that didSave was forgetting to log status, which is a
less likely mistake if we have a struct).

But to do that, I first need to make the name `DaemonStatus` available for the plain
dataclass.

Reviewed By: grievejia

Differential Revision: D43243255

fbshipit-source-id: c5ac7ac8db25c5498bfb473f41284bd856fddc6a