Always get time since ready, status together
Summary:
We always want these two bits of data at the same time, let's make sure
the code reflects that by always doing them on consecutive lines.

This is a trivial refactor that will make an upcoming diff (creating a status
dataclass to encapsulate both of these) easier to write and read.

Reviewed By: grievejia

Differential Revision: D43244194

fbshipit-source-id: 1b86046b728c0b43fbef3934754bf94a5f8160bd