Use DaemonStatus to encapsulate time + connection status
Summary:
We always want both the connection status before and information about
how long it has been since the server is ready; this combination of information
makes it much easier to judge how big a problem availability is, because
- we can't decide whether incremental status is bad or not *at all* without
  a timer, since it's expected that we'll be briefly busy
- even for other statuses like starting, it's very helpful to have a record
  of how long start took; this, for example, can help us distinguish between
  an availability problem due primarily to long start times versus a problem
  due primarilily to frequent deamon (or persistent) restarts

Putting these fields in the same dataclass makes it easier to ensure we always
do the right thing, and that the logging is consistent across all telemetry
events. It's worth noting that I caught another bug here: we weren't logging status
in the telemetry for type coverage; this isn't a big deal today but it could
be one if we ever start seeing heavy use of expression-level coverage.

Reviewed By: grievejia

Differential Revision: D43244195

fbshipit-source-id: d5ff1b0a39e2b09e3873a3131373a7f466c77d5d