Remove warning from dataclasses_json in code_navigation_request
Summary:
There's an [issue](https://github.com/lidatong/dataclasses-json/issues/62) with dataclasses_json supporting `from __future__ import annotations`. This diff:
- Removes `from __future__ import annotations`
- adds `mm_field` argument to `dataclasses.field`

The second bullet point seems like it wouldn't be necessary but it was the only thing that got the error message to go away. It seems like it fixed the false negative on `ValidationError`s as well.

Alternative options to this would be:
- remove reliance on dataclasses_json
This is a long term solution that we might want especially when newer python versions implicitly include `from __future__ import annotations`. It requires reworking the entirety of `code_navigation_request` so I am leaving it for the future.
- Fully rely on `dataclasses_json`
We can't do this that well because we need to override certain json encodings because of the protocol.

Reviewed By: stroxler, grievejia

Differential Revision: D43201933

fbshipit-source-id: 29e769a4a3c2c50c031a6602c7ff786593978b1b