Use early return in process hover logic
Summary:
In our python code early return is the usual idiom. I don't actually
like it much, but given the need for refinement and not indenting too much
it is pragmatic; more importantly, it is already our convention especially
in the various language server handler functions.

The hover function was using an if/else instead. This is harder to follow
because it is different from the other handlers, plus the mismatched
indentation keeps making easy refactors slightly more annoying because I
have to pause and adjust indentation.

This diff fixes it.

Reviewed By: jasontatton

Differential Revision: D43258773

fbshipit-source-id: 0033426f9a662df161aa13baf6f8d380328a0fc4