Fix pyre rage for servers with extensions
Summary: extensions were not able to be encoded in json. this diff implements that ability to be displayed in pyre rage. In this process, I attempted to redo the dataclass to use `dataclasses_json` but I wasn't able to switch to that while maintaining the invalid configuration exception we currently have.

Reviewed By: stroxler

Differential Revision: D43365473

fbshipit-source-id: 696480d5e40c0d5dfb1e4f59e856bdec81870987