Populate location info in from imports
Summary:
See previous diff for more context. This diff populates the locatio info by estimating based on characters.

Note: The following cases do not work with this approach.
```
from        x import y
```
(added test)
```
from\
x import y
```
(Pyre can't seem to parse it but is valid python)

This info is still not added to locationBasedLookup, which will come in the next diff.

Reviewed By: stroxler

Differential Revision: D43402052

fbshipit-source-id: 77a1e5f33ad216962ce031025c10e24fd0b3f485