Add location information to from import in AST
Summary:
bruhmoment
(macro stolen from D31134521 (https://github.com/facebook/pyre-check/commit/86fab087805b5760e652b6e2abff1eb721274ff2))

Pyre's go-to definition is failing on the x in `from x import y`. This is because the cpython AST does not contain location info for that token.

As a workaround, we will be adding information about this to our own AST. We know the `x` starts 5 characters from the beginning of the line, so we can use that.

This diff modifies the AST with this information - the next diff will populate it.

Reviewed By: stroxler

Differential Revision: D43397107

fbshipit-source-id: 16838e959aa589c210014076b7f79a621ebf18a2