Handle global leak detection for inner function statement
Summary:
This diff stack implements a global leak detection framework on top of Pyre.
We would like the ability to check for global leaks that occur in nested `def` statements. We dont need to do anything additional for this statement case. The statements within try are handled by forward. This diff adds various test cases for `def`.

Reviewed By: connernilsen

Differential Revision: D43079640

fbshipit-source-id: f6ca5f571e146892709dc353dc095d04a2b2e19f