Find flows into implicit string sinks inside format strings
Summary:
This diff only finds multi-source issues where both flows are in the same
function. In other words, this diff does not yet propagate the triggered sinks,
which will be addressed in a follow-up.

Implementation-wise, this diff:
1. Create module `TaintConfiguration.StringOperationPartialSinks` to group
operations related with storing and creating partial sinks for string
operations (such as format strings).
2. Check multi-source flows into implicit string sinks that are introduced onto
expressions used inside format strings and string functions (str.__add__,
str.__mod__, and str.format) when at least one of the arguments is a string
literal.

Reviewed By: arthaud

Differential Revision: D43101771

fbshipit-source-id: 2032195363d95da2eadc1ecf24bf345cd1b25380