Go-to definition for keyword arguments
Summary:
We currently do not support go-to definition on keyword arguments. In an ideal world, clicking an argument x would take you to the parameter x in the definition. For example:
```
def foo(x: int):
        ^ b
    ...
foo(x=5)
    ^ a
```
Clicking `a` should bring you to `b`. This is complicated, requiring knowledge of the local context within foo. For now, this diff brings you to the definition of `foo` (not the exact parameter) by adding a new `callee` in the visit record. This may not be the best approach; I'm open to suggestions.

The next diff (D43517176) attempts to use this information to look up the definition and call `look_up_local_definition` however we do not get the define_name making it not work successfully. If this apprach makes sense to others, I can invest more time in the next diff.

Reviewed By: stroxler

Differential Revision: D43512773

fbshipit-source-id: 405b452271d4008578a1ecca1dd944560d186c15