Use correct locations for keyword arguments
Summary:
In both of our parsers, we currently claim the location of keyword arguments in a function call is the location of the entire call. This is incorrect, and was done previously to align with the cpython AST.

With code navigation features coming out, we will need this location to be correct to enable go-to definition on these symbols. This diff corrects the locations in our two current parsers. The next diffs will continue in supporting go-to def on these symbols.

Reviewed By: grievejia

Differential Revision: D43512774

fbshipit-source-id: 0f1fe463885b175344d00f3de2de97fe59495ff0