Move `log_lsp_event` up to `client/`
Summary:
I'm not very happy with the current remote logging code layout, we're
pretty inconsistent about how we handle it (sometimes we assume that
the `logger` command is non-None and force callers to handle it,
sometimes we take an arbitrary `Configuration` and check whether the
logger is `None`, sometimes we take an optional `RemoteLogging` value
which is even weirder since that's part of the `BackendArguments`
code).

But refactoring that is going to take some thought, in the meantime
I'd like to just keep migrating code out of `commands`, in order to
unblock moving the core language server components into `language_server`.

Reviewed By: grievejia

Differential Revision: D43563909

fbshipit-source-id: cd9304beb0f0e084701de47dd6cadb934d4df184