Use Name location in cfg identifier
Summary:
The name location has been incorrect. Because of that, we had been creating cfgs with the location of the entire handler. When going to definition, this has been incorrectly leading to the error expression, not the definition.

With the new correct locations, this diff makes go-to definition correct for exceptions.

Reviewed By: grievejia

Differential Revision: D43674086

fbshipit-source-id: e96c2a5592e15637023eff0daf90338ad5071dab