Docstring + function rename in typeshed pull
Summary:
I was reading through the typeshed download + patch code to understand
what it would take to make typeshed a directory rather than a zipfile,
and I figured a docstring would be good. I also decided to rename the
output-writing function to be explicitly about writing to a zipfile because
that will make it easy for me to play with writing to a directory instead.

Reviewed By: kinto0

Differential Revision: D43664087

fbshipit-source-id: a1c3e316a652f2575fa7e5c24eb09fbb78207d9b