Add integration tests to demonstrate false negatives in detecting implicit string literal sinks
Summary:
When an argument of a string formatting call (e.g., `str.__mod__`) is a string
literal, we currently do not attach implicit sinks.

The test case in `implicit_sources.py` is not about false negative, but is
similar -- an argument of a string formatting call is a string literal.

Reviewed By: alexkassil

Differential Revision: D43700703

fbshipit-source-id: 63a5d15514a0d430488968c7debbabb506e7e9ff