Add a test illustrating that Pyre mis-handles descriptor overrides
Summary:
It turns out that Pyre is not correctly checking equivalent attribute lookups when doing
override compatibility checks - it complains if we override a descriptor where the
class and instance lookups are different, because it checks the instance lookup on the
child versus the class lookup on the parent.

This commit adds a test demonstrating the problem.

Reviewed By: grievejia

Differential Revision: D43722605

fbshipit-source-id: 0d8d597611cc9f2b7f021d81d1e35b34e5f9d313