Use class lookups for override compatiblity (hack)
Summary:
This commit changes override compatiblity checks to always use
class-based lookups.

The reason for this is that `GlobalResolution.overrides` always uses
class-baded lookups (https://fburl.com/code/3cqbl43p), probably because
that happens to work reasonably well given that today Pyre isn't very
good at distinguishing class vs instance attributes anyway, and we do
want to have sane behavior on static methods / class methods which
are very common.

But, the override check was using this class-based lookup on the parent
and verifying compatilility against an *instance* lookup on the child. That's
just wrong! In the case of a descriptor where instance and class access don't
behave the same, it will produce false positives.

For now I just want to do the minimal work to fix our false positive, and
the quick and dirty solution is to always use class lookups. That fixes the
false positive, although it still means that for a descriptor that uses different
types on class vs instance access we'll get an incorrect check (to be clear this
isn't really a regression, since that problem already exists for class
lookups - the new behavior is strictly better it's just not perfect).

I filed a task to do more thorough checks. It's unclear how urgent this is;
descriptors with overloads for class vs instance attributes probably aren't
that common so avoiding false positives is probably a decent start.

Reviewed By: grievejia

Differential Revision: D43722606

fbshipit-source-id: a485395f5daf58f9465ec87b1ade25832a37b9aa