Hardcode models with partial sinks for string formatting functions in the forward analysis
Summary:
Such as `str.__add__`, `str.__mod__`, `str.format`.

This diff also removes the combined source rules and the user models for the
above functions in the integration tests. Such removal causes changes in the
*.models files. These changes are not material except for the model of
`DatabaseSchemaEditor.string_operations` (see below for more details) -- the
fact that the models remain unchanged even after deleting the user-models for
the string format functions means the implementation is correct.

Note, we lose the model for
`multiple_sources_string_operations.DatabaseSchemaEditor.string_operations`
because this is a sink model, which is inferred by the backward analysis but we
have not yet updated the backward analysis. The next diff will reintroduce this
model.

Reviewed By: alexkassil

Differential Revision: D43555107

fbshipit-source-id: ee91658463bfee8d19f35fceb73ab0e2b1165cda